Update graphql version, dependabot, file org
