+++ .github/dependabot.yml
@@ -7,3 +7,8 @@ updates:
     directory: "/" # Location of package manifests
     schedule:
       interval: "daily"
+    ignore:
+      - dependency-name: "graphql"
+        # Currently the graphql module seems to break on Codesandbox when 
+        # over version 16, so we'll keep it on 15.x until this is resolved.
+        # See issue apollo-client/#9943.
