+++ package.json
@@ -5,7 +5,7 @@
   "license": "MIT",
   "dependencies": {
     "@apollo/client": "^3.6.8",
-    "graphql": "^16.4.0",
+    "graphql": "^15.6.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
